refactor: move permissions to core, services to datasets app, clean up views and tasks
