+++ apps/datasets/serializers.py
@@ -1,7 +1,5 @@
 from rest_framework import serializers
 
-from services.dataset_utils import get_file_type
-
 from .models import Dataset
 
 
@@ -16,9 +14,3 @@ class DatasetSerializer(serializers.ModelSerializer):
         if ext not in ["csv", "xlsx", "xls", "xlsm"]:
             raise serializers.ValidationError("Поддерживаются только CSV и Excel файлы")
         return value
-
-    def create(self, validated_data):
-        file_type = get_file_type(validated_data["file"].name)
-        validated_data["file_type"] = file_type
-        validated_data["user"] = self.context["request"].user
-        return Dataset.objects.create(**validated_data)
