+++ apps/datasets/views.py
@@ -1,10 +1,11 @@
 from rest_framework import viewsets
 from rest_framework.permissions import IsAuthenticated
 
-from apps.users.permissions import IsOwner
+from apps.core.permissions import IsOwner
 
 from .models import Dataset
 from .serializers import DatasetSerializer
+from .services import get_file_type
 from .tasks import process_dataset
 
 
@@ -16,5 +17,6 @@ class DatasetViewSet(viewsets.ModelViewSet):
         return Dataset.objects.filter(user=self.request.user)
 
     def perform_create(self, serializer):
-        dataset = serializer.save()
+        file_type = get_file_type(serializer.validated_data["file"].name)
+        dataset = serializer.save(user=self.request.user, file_type=file_type)
         process_dataset.delay(dataset.pk)
